Managed to fix the picking to be at the right at the right relative position

+++ engine/render/frames/flatgeometrylitpass.cpp
@@ -67,7 +67,7 @@ void FlatGeometryLitPass::Setup()
 
 	glGenTextures(1, &this->uvBuffer);
 	glBindTexture(GL_TEXTURE_2D, this->uvBuffer);
-	glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB16F, res.x, res.y, 0, GL_RGB, GL_FLOAT, NULL);
+	glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB32F, res.x, res.y, 0, GL_RGB, GL_FLOAT, NULL);
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
